pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vendor-atlas-mcp"
version = "0.1.0"
description = "Vendor Atlas MCP server and consumer dashboard for vendor market discovery."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Vendor Atlas", email = "support@example.com"},
]

dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "python-dotenv>=1.0",
    "stripe>=7.0",
    "httpx>=0.27",
    "itsdangerous>=2.2",
]

[project.optional-dependencies]
reports = [
    "jinja2>=3.1",
    "weasyprint>=60.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[project.scripts]
vendor-atlas-mcp = "server:main"

[tool.setuptools]
py-modules = [
    "billing",
    "config",
    "server",
    "storage_events",
    "storage_markets",
    "storage_users",
]

[tool.setuptools.packages.find]
include = ["middleware*", "tools*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
asyncio_mode = "auto"
testpaths = ["tests"]
